+++ src/backend/ingestion/framework-detection.lambda-shared.ts
